chore: set ALLOW ORIGIN HEADER on identifier
This makes CORS work when e.g. reaching the SPARQL endpoint from a Yasgui instance (like [the one hosted by redpencil](https://yasgui.tools.redpencil.io/)).

We could also add this header ONLY to the SPARQL endpoint in the dispatcher but I'm not sure that that's valuable (i.e. we're not doing anything wrong by setting it for the whole app, I think).